New Feature: Get directions for Google Places API
Routing class is now AbstractRouting class. For backward compatibility,
Routing class is working as it was. Google Places API results now can
be used for getting directions using PlaceRouting class.
Updating
